getting all the patients info in postman after get request

+++ src/main/java/com/Om/DentalClinic/repository/PatientInfoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Om.DentalClinic.model.PatientInfo;
 
-public interface PatientInfoRepository extends JpaRepository<PatientInfo,Integer> {
+public interface PatientInfoRepository extends JpaRepository<PatientInfo,Long> {
 
 }
